chore: Error.cpp, Reply.cpp 합침, ERROR 답장용 함수 추가
