+++ Makefile
@@ -6,7 +6,7 @@
 #    By: jiychoi <jiychoi@student.42seoul.kr>       +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/12/28 13:47:07 by jiychoi           #+#    #+#              #
-#    Updated: 2023/01/02 22:07:48 by jiychoi          ###   ########.fr        #
+#    Updated: 2023/01/04 12:52:08 by jiychoi          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -15,7 +15,7 @@ CXXFLAGS	=	-Wall -Wextra -Werror -std=c++98
 SRCS		=	src/main.cpp \
 				src/Server.cpp src/User.cpp \
 				src/Command.cpp src/CommandChannel.cpp src/CommandAuth.cpp \
-				src/Error.cpp src/utils.cpp src/Reply.cpp \
+				src/utils.cpp src/Reply.cpp \
 				src/Channel.cpp
 OBJS		=	$(SRCS:.cpp=.o)
 TARGET		=	ircserv
